requirements.txt
PyQt6>=6.6.0
mss>=9.0.1
Pillow>=10.0.0
pywin32>=306; platform_system == "Windows"
pyautogui>=0.9.54
imagehash>=4.3.1
numpy>=1.24.0
keyboard>=0.13.5
